Fix mongo start in KOS qemu: change compiler -march option to armv8-a
Trying to run mongod in KOS qemu causes the program to crash with an
"Unhandled Generic Exception" message and a stack trace.

Investigating the crash log it was discovered that the problem is
connected with fetch_add() function (atomic add). Investigating the
disassembler listing it was found that the problematic instruction is
LDADDAL. Support for this instruction  was introduced in ARMv8.1.

For *aarch64* target architecture, the SConstruct script sets the -march
compiler option to armv8.2-a. KOS qemu does not support armv8.2-a. The
most supported is armv8-a.

Signed-off-by: Mikhail Zhvakin <mikhail.zhvakin@auriga.com>