Fix usage of __wt_rdtsc()
KOS aarch64: Attempting to read the CNTVCT_EL0 register using the MSR
instruction throws an exception (causing the program to crash).
Testing reading this register on different versions of QEMU-aarch64 with
the linux kernel was successful. But with KOS kernel read register
failed with the same exception.
It looks like the KOS kernel is not correctly setting up the CPU
registers to access the CNTVCT_EL0 register from the EL0 (Exception
Level 0, i.e. Application level). Or does not have a corresponding
handler.

For KOS, the epoch time will be used instead of the TSC value.

TODO: delete __wt_rdtsc() and __global_calibrate_ticks() W/A for KOS.

Signed-off-by: Mikhail Zhvakin <mikhail.zhvakin@auriga.com>