Add W/A so gethostname() doesn't return an empty string
The file 'hostname' exists in ROMFS whose mount point is the /etc
directory (VfsNet entity). So the /etc/hostname path is present in the
VFS, but gethostname() returns an empty string. This causes a lot of
error messages in the log such as "cannot get hostname".

The main purpose of this W/A is to clear the log of these messages.

TODO: remove the W/A when gethostname() works correctly.

Signed-off-by: Mikhail Zhvakin <mikhail.zhvakin@auriga.com>